dpkg: Change --command-fd file descriptor to an int
Instead of using a long with a cast for the format string, use an int
which is what the dpkg_options_parse_arg_int() function parses and
checks for as the allowed value range.

Changelog: internal

+++ src/main/main.c
@@ -858,7 +858,7 @@ commandfd(const char *const *argv)
 	const char **newargs = NULL, **endargs;
 	char *ptr, *endptr;
 	FILE *in;
-	long infd;
+	int infd;
 	int ret = 0;
 	int c, lno, i;
 	bool skipchar;
@@ -871,7 +871,7 @@ commandfd(const char *const *argv)
 	infd = dpkg_options_parse_arg_int(cipaction, pipein);
 	in = fdopen(infd, "r");
 	if (in == NULL)
-		ohshite(_("cannot open '%i' for stream"), (int)infd);
+		ohshite(_("cannot open '%i' for stream"), infd);
 
 	lno = 0;
